implement a fix for the delete data function.
The delete data function was attempting to print the whole object to the
console but that is not allowed when constructing Objects. I first tried
to overload the operator'<<' as a member function but the compiler did
not like this. ChatGPT suggested I try overloading it as a global
function. It worked but I am not totally sure why. I think it has
something to do with needing to use all member variables maybe. But once
you move it to a global function that need goes away.